inprogress

+++ FastPricingEngine/src/com/upupconsultant/common/data/Dao.java
@@ -6,7 +6,7 @@ import com.upupconsultant.pricing.model.PricingRule;
 import com.upupconsultant.pricing.model.PricingRuleCondition;
 
 public class Dao {
-	public List<PricingRule> findPricingRule(int providerId){
+	public List<PricingRule> findPricingRule(long providerId){
 		return null;
 	}
 	public List<PricingRuleCondition> findPricingRuleCondition(long ruleid){
